requirements.txt
asgiref==3.11.1
bcrypt==5.0.0
cachetools==7.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
ecdsa==0.19.1
Faker==40.4.0
google-auth==2.48.0
hiredis==3.3.0
idna==3.11
pillow==12.1.1
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
python-dotenv==1.2.1
python-jose==3.5.0
redis==7.2.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
